Refactor import paths and component names in course-related files
