+++ src/store/course.ts
@@ -19,7 +19,7 @@ import { createApplicationError } from "@/utils/create-application-error.factory
 import { Action, Module, Mutation, VuexModule } from "vuex-module-decorators";
 import { BusinessError } from "./types/commons";
 import { HttpStatusCode } from "./types/http-status-code.enum";
-import { Course } from "./types/room";
+import { Course } from "./types/course";
 
 @Module({
 	name: "courseModule",
